requirements.txt
tweepy
praw
beautifulsoup4
nltk
spacy
textblob
transformers
ccxt
flask
python-dotenv